Fix insightface installation in Docker container (#333)
This commit resolves an issue encountered while installing the insightface package in a Docker container. The installation process was failing due to missing Python development headers, specifically the Python.h file.

The Dockerfile has been updated to install the python3.11-dev package, which provides the necessary header files and development libraries for building Python extensions. With this change, the insightface package can now be successfully installed in the Docker container.