Add 16bits color depth support (#255)
* Add 16bits color depth support

Pillow doesn't support natively so OpenCV was used when processing images requested as 16bits.
That image is later converted to a Pillow image when passing it to the comfy ProgressBar.
When requesting an image as 8bits RGB (the default for all formats), or when requesting a format other than PNG, the original image processing is used.

* Add 32bit color depth mode and Tiff support

* Remove 32bit and tiff support

The changes to support 32bit, and tiff to actually make use of true 32bit, were pretty significant and would require separate consideration. Extracting these changes out into a separate branch for a later time.

* Remove unecessary mode check

* Don't convert bits in C#

Use the generated image base on user params to determine how many bits to encode the png with

* Revert accidental whitespace removal

In vscode, trailing whitespaces are by default removed. To prevent that behavior this setting is needed: `files.trimTrailingWhitespace: false`

* Adress feedback

* Fix syntax and address feeback